+++ lib/Perl/Tidy/IOScalarArray.pm
@@ -14,7 +14,7 @@ package Perl::Tidy::IOScalarArray;
 use strict;
 use warnings;
 use Carp;
-our $VERSION = '20190915';
+our $VERSION = '20190915.01';
 
 sub new {
     my ( $package, $rarray, $mode ) = @_;