+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## 2019 06 01.01
 
+    - fixed issue git#13, needless trailing whitespace in error message
+
     - fixed issue git#9: if the -ce (--cuddled-else) flag is used,
       do not try to form new one line blocks for a block type 
       specified with -cbl, particularly map, sort, grep