+++ CHANGES.md
@@ -2,6 +2,14 @@
 
 ## 2019 06 01.01
 
+    - fixed for RT#129850; concerning a space between a closing block brace and
+      opening bracket or brace, as occurs before the '[' in this line:
+
+       my @addunix = map { File::Spec::Unix->catfile( @ROOT, @$_ ) } ['b'];
+
+      Formerly, any space was removed. Now it is optional, and the output will
+      follow the input.  
+
     - fixed issue git#13, needless trailing whitespace in error message
 
     - fixed issue git#9: if the -ce (--cuddled-else) flag is used,